Better delineate speech title and speaker's name
I find it difficult to see where the speech title ends and speakers name begins.  This change simply adds a `-` between them, but more complex solutions could be made (colour or layout?).

It appears as though the "Meet + Greet" and "Pizza" entries are just a `speaker` with no `speaker` so I just removed the whole bit if it's not there.